Remove note to process inner file(s) piecewise
Initially, I thought to only process files piecewise, so that the full
sequence of bytes pertaining to the files inside the archive(s) are
never in one piece.

But the `zip` crate and the `sha2` crate will both probably have them
in memory anyway, and I don't think there's much of a risk doing so,
as long as we're not executing the file.